Uploading corrected version of homework
I am uploading my homework again with the minor corrections (needed to address the previous "float not subscriptable" error) I made with help from the discussion section. The corrections are in ckrk_mv_core Ln 79 (capitalized a to A) and ckrk_mv_ad Ln 49 (capitalized delta to Delta to properly update it in the while loop).